chore: améliorations à l'obtention d'évènements Outlook
